To pR without maps and trip components
