+++ packages/app/components/PackOptions/index.tsx
@@ -1,15 +1,17 @@
-import { useState } from 'react';
 import { View } from 'react-native';
-import { useDispatch } from 'react-redux';
-import { useRouter } from 'app/hooks/router';
 import { RStack, RIconButton, BaseModal } from '@packrat/ui';
 import { Entypo } from '@expo/vector-icons';
+import { useModalState } from './useModalState';
 export const PackOptions = ({ Edit, Delete, Ignore }) => {
-  const dispatch = useDispatch();
-  const router = useRouter();
+  const { isModalOpen, openModal, closeModal } = useModalState();
+
   return (
     <View>
-      <BaseModal triggerComponent={<MenuTriggerComponent />}>
+      <BaseModal
+        isOpen={isModalOpen}
+        onClose={closeModal}
+        triggerComponent={<MenuTriggerComponent openModal={openModal} />}
+      >
         <RStack style={{ flexDirection: 'row', alignItems: 'center' }}>
           {Edit} Edit{' '}
         </RStack>
@@ -30,7 +32,7 @@ export const PackOptions = ({ Edit, Delete, Ignore }) => {
   );
 };
 
-const MenuTriggerComponent = ({ setIsModalOpen }) => {
+const MenuTriggerComponent = ({ openModal }) => {
   return (
     <View
       style={{
@@ -43,9 +45,7 @@ const MenuTriggerComponent = ({ setIsModalOpen }) => {
       <RIconButton
         backgroundColor="transparent"
         icon={<Entypo name="dots-three-horizontal" size={24} color="black" />}
-        onPress={() => {
-          setIsModalOpen(true);
-        }}
+        onPress={openModal}
       />
     </View>
   );
